refactor: extract get_header to utils/auth_helper.py.
- Move get_header to utils/auth_helper.py for reusability.
- Remove delete_event function to focus on implementing get_event function in this branch.
